Add Deposit method to Table
Adding a deposit method that stores the deposit made by the user to a data field. Passed tests, and have written failing test for Withdraw method

+++ Bank/src/Bank.cs
@@ -11,7 +11,7 @@ namespace src
              return _initialBalance;
          }
 
-         public double Deposit(double add)
+         public double Deposit(int add)
          {
              if (add <= 0)
              {
